Add Support for XML Importer
Run the Reflection field on Post Create / Edit within the XML Importer by registering for the delegate.

For those seeking to use this functionality it is of utmost importance that in the XML Importer you create a 'dummy' entry such as `concat("","")`to create the initial row in the database. Reflection field only does an update on the field, thus if no entry is found in the DB this will remain blank.